feat(silver): correct FUELHH settlement identity from vendor start instant (v0.21-F sp48)
Derive settlement_date/settlement_period/timestamp_utc from the Elexon
startTime instant (with legacy startTimeOfHalfHrPeriod alias support and a
counted label-based fallback) instead of trusting the vendor's
settlementDate/settlementPeriod label alone. Fixes last-period-of-day
mislabels that collided distinct settlement periods onto one entity key
across DST boundaries. Bumps DATASET_VERSION to 2.0.0 and threads the new
start_time_fallback_count into PipelineRunTracker per-run reporting
(REVIEW-PLAN-1 M-1, M-2). Includes the ADR-031 wording rider and the CLI
status duration double-rounding rider (both T0).

Verified via a persisted temp-root rebuild across all four named probe
windows (Sept 2021, Oct/Nov 2021 DST, Mar 2022 DST, Jul 2022): zero
timestamp mismatches, zero generation-conflicting keys, complete
46/48/50-period calendars, zero start_time_fallback_count, dataset_version
2.0.0 throughout. Source bronze bytes unchanged (SHA-256 verified).

+++ src/gridflow/cli.py
@@ -599,7 +599,7 @@ def status() -> None:
     try:
         result = con.sql("""
             SELECT source, dataset, operation, status,
-                   rows_out, ROUND(duration_seconds, 1) as duration_s
+                   rows_out, ROUND(duration_seconds::DOUBLE, 1) as duration_s
             FROM pipeline_runs
             WHERE started_at > now() - INTERVAL '24 hours'
             ORDER BY started_at DESC
